rename: Journal컴포넌트 이름 변경
- Journal -> JournalsSection 으로 변경
