+++ src/components/journalsSection/JournalsSection.module.scss
